Fix for numberformat exception on large int's changed them to longs

+++ src/mgi/frontend/datamodel/sort/SmartAlphaComparator.java
@@ -19,16 +19,13 @@ import java.util.Set;
 	 * author kstone
 	 * 
 	 */
-public class SmartAlphaComparator<T> implements Comparator<T>
-{
+public class SmartAlphaComparator<T> implements Comparator<T> {
 	/* The following algorithm was taken from the symbolsort module in lib python */
 	
 	private Set<Character> digits = new HashSet<Character>(Arrays.asList('0','1','2','3','4','5','6','7','8','9'));
-	private ArrayList<Object> splitter (String s)
-	{
+	private ArrayList<Object> splitter (String s) {
 		ArrayList<Object> splits = new ArrayList<Object>();
-		if(s==null || s=="")
-		{
+		if(s==null || s=="") {
 			splits.add(s);
 			return splits;
 		}
@@ -37,28 +34,20 @@ public class SmartAlphaComparator<T> implements Comparator<T>
 
 		boolean inDigits = digits.contains(s.charAt(0));
 
-		for (int i =0; i < s.length();i++)
-		{
-			if (digits.contains(s.charAt(i)) != inDigits)
-			{
-				if (inDigits)
-				{
-					splits.add((Integer)Integer.parseInt(s.substring(last, i)));
-				}
-				else
-				{
+		for(int i = 0; i < s.length(); i++) {
+			if (digits.contains(s.charAt(i)) != inDigits) {
+				if (inDigits) {
+					splits.add(Long.parseLong(s.substring(last, i)));
+				} else {
 					splits.add(s.substring(last,i));
 				}
 				last = i;
-				inDigits = ! inDigits;
+				inDigits = !inDigits;
 			}
 		}
-		if (inDigits)
-		{
-			splits.add(Integer.parseInt(s.substring(last)));
-		}
-		else
-		{
+		if (inDigits) {
+			splits.add(Long.parseLong(s.substring(last)));
+		} else {
 			splits.add(s.substring(last));
 		}
 		//return items;
@@ -68,42 +57,34 @@ public class SmartAlphaComparator<T> implements Comparator<T>
 	/*
 	 * This method should be overriden to be useful
 	 */
-	public int compare(T o1, T o2)
-	{
+	public int compare(T o1, T o2) {
 		return compare(o1.toString(),o2.toString());
 	}
 	
 	/* these values must really be Strings, but we want to keep the 
 	 * class level type as Object to allow for flexible subclassing
 	 */
-	public int compare(String o1, String o2) 
-	{
+	public int compare(String o1, String o2)  {
 		String string1 = (String) o1;
 		String string2 = (String) o2;
 		// are these both objects that we can compare?
-		if(o1!=null && o2!=null)
-		{
+		if(o1 != null && o2 != null) {
 			// split them into parts first
 			ArrayList<Object> s1 = splitter(string1);
 			ArrayList<Object> s2 = splitter(string2);
 			
 			// go through each part until we find a difference
-			for (int i=0; i< Math.min(s1.size(), s2.size());i++)
-			{
+			for (int i = 0; i< Math.min(s1.size(), s2.size()); i++) {
 				Object ob1 = s1.get(i);
 				Object ob2 = s2.get(i);
 				//check instance
-				if(ob1.getClass() != ob2.getClass())
-				{
+				if(ob1.getClass() != ob2.getClass()) {
 					// numbers go first
 					if(ob1 instanceof Integer) return -1;
 					else return 1;
-				}
-				else
-				{
+				} else {
 					// continue on if they are equal
-					if (!ob1.equals(ob2))
-					{
+					if (!ob1.equals(ob2)) {
 						// just cast the correct type and return the compare functions for Integer or String
 						if(ob1 instanceof Integer) return ((Integer)ob1).compareTo((Integer)ob2);							
 						else return ((String)ob1).compareToIgnoreCase((String)ob2);
@@ -115,12 +96,9 @@ public class SmartAlphaComparator<T> implements Comparator<T>
 			else if (s1.size() > s2.size()) return 1;
 			// well, they must be equal then.
 			return 0;
-		}
-		else
-		{
-			if (o2==null)
-			{
-				if(o1==null) return 0;
+		} else {
+			if (o2 == null) {
+				if(o1 == null) return 0;
 				return -1;
 			}
 			else return 1;
